add game.js class
